week13/requirements.txt
langchain
langchain-community
langchain-ollama
arize-phoenix
openinference-instrumentation-langchain
opentelemetry-api
opentelemetry-sdk
pandas
pytest
